requirements.txt
azure-cognitiveservices-speech==1.41.1
Pillow==10.4.0
# if you use rich logging or extras, add them here
